WIP: Migrate RemarkableBackend, split out web server
