--- src/remarkable/mod.rs
@@ -1,3 +1,0 @@
-mod remarkable_app;
-
-pub use remarkable_app::RemarkableApp;
